Release v1.0.6
This commit contains changes that improve the SSO connection, especially the part where the SDK exchanges a client token for an Octopus Community JWT.
In certain cases, the client token was asked too many times and was sometimes stored to be used later, even if it had expired. To improve it, we do not store it anymore, instead we use it directly.

It also contains small fixes such as localization fixes and author's name UI fixes.