Changed assets usage to custom generated ones
To add support for Cocoapods in static library mode, where assets are not generating the Asset extension (even if the xcassets file has this option enabled), this commit changes the way we are using assets in the project in order to work when used with SPM or Cocoapods.

To do that, and to keep using assets (colors and images) by their symbols and not their explicit names, we:
- disabled the xcassets extension generation
- created a script that generates the swift asset symbols file and the extensions of main structs (UIColor, Color, UIImage) to directly use the symbols
- changed the usage in the code (normally, small impact but sometimes as extensions were conflicting, we had to add `res:` or `Gen` as prefix

Note: this commit is still not working in Cocoapods because the db and assets files are not found in their bundle.
This is because the following code:
```
private class ResourceBundleClass {}
private let resourceBundle = Foundation.Bundle(for: ResourceBundleClass.self)
```
is returning the main bundle even if ResourceBundleClass is placed in the module OctopusUI.
When we look at the build folder, we correctly have a bundle called OctopusUI and another OctopusCore.